requirements.txt
requests==2.31.0
paramiko>=3.4.0
cryptography>=42.0.0
python-dotenv==1.0.0
